requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0

# ── PDF & DOCX parsing ─────────────────────────────────────────────────────────
pdfplumber==0.11.0
PyMuPDF==1.24.3
python-docx==1.1.2

# ── LLM / AI ──────────────────────────────────────────────────────────────────
google-generativeai==0.8.2
sentence-transformers==3.0.0

# ── Database & cache ──────────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.6

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
httpx==0.27.0
python-dateutil==2.9.0
dateparser==1.2.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
